EMQUAL, an anonymous working recruitment website
With the ongoing problem of gender employment and pay gap, questions came into our minds. What if the only factor to be judged upon job applications is skills? What if all genders, races, and nationalities are equal? We, ATLAS(T), wanted to be a part of creating a world where employment is judged based on the skill, and only the skill, of the employees. This inspired us to create ‘EMQUAL’ an anonymous work recruitment website, where gender, race, nationality, personal information, and social status does not take advantage of one's skills and commitment.